order-service/requirements.txt
django
djangorestframework
requests
pika
PyMySQL


cryptography
